+++ libs/cardano-ledger-core/CHANGELOG.md
@@ -1,5 +1,9 @@
 # Version history for `cardano-ledger-core`
 
+## 1.15.0.1
+
+*
+
 ## 1.15.0.0
 
 * Add `Mismatch` type to clarify predicate-failures reporting supplied and expected values. #4649